1dmodel/requirement.txt
pytorch
pytorch-transformers
tqdm
numpy
